+++ app/src/main/java/eu/kanade/tachiyomi/ui/setting/SettingsBackupController.kt
@@ -124,7 +124,7 @@ class SettingsBackupController : SettingsController() {
 
                 preferences.backupsDirectory().changes()
                     .onEach { path ->
-                        val dir = UniFile.fromUri(context, path.toUri())
+                        val dir = UniFile.fromUri(context, path.toUri())!!
                         val filePath = dir.filePath
                         summary = if (filePath != null) "$filePath/automatic" else "Invalid directory: ${dir.uri}"
                     }